API Reference

DitherPunk.BinaryDitherMetric

DitherPunk.ClosestColor

DitherPunk.ConstantThreshold

DitherPunk.ErrorDiffusion

DitherPunk.ErrorDiffusion

DitherPunk.OrderedDither

DitherPunk.WhiteNoiseThreshold

DitherPunk.BalancedCenteredPoint

DitherPunk.Bayer

DitherPunk.CentralWhitePoint

DitherPunk.ClusteredDots

DitherPunk.Rhombus

DitherPunk.bayer_matrix

DitherPunk.closest_color

DitherPunk.dither

DitherPunk.dither!

DitherPunk.gradient_image

DitherPunk.linear2srgb

DitherPunk.srgb2linear

DitherPunk.test_on_gradient

DitherPunk.upscale

Functions

DitherPunk.BinaryDitherMetric — Type

Define custom color difference metric with linear distances between grayscale values.

DitherPunk.ClosestColor — Type

Simplest form of image quantization by turning each pixel to the closest one in the provided color palette cs. Technically this not a dithering algorithm as the quatization error is not "randomized".

DitherPunk.ConstantThreshold — Type

ConstantThreshold(threshold)

Use a constant threshold map. Defaults to 0.5 if threshold isn't specified.

DitherPunk.ErrorDiffusion — Method

Binary error diffusion.

DitherPunk.ErrorDiffusion — Method

Error diffusion for general color schemes cs.

DitherPunk.OrderedDither — Method

dither(img::AbstractMatrix, alg::AbstractOrderedDither; to_linear=false, invert_map=false)

Ordered dithering using threshold maps. Takes a grayscale image img and a normalized threshold matrix mat. The threshold matrix is repeatedly tiled to match the size of img and is then applied as a per-pixel threshold map. Optionally, this final threshold map can be inverted by selecting invert_map=true.

DitherPunk.WhiteNoiseThreshold — Type

WhiteNoiseThreshold()

Use white noise as a threshold map.

DitherPunk.BalancedCenteredPoint — Method

BalancedCenteredPoint()

Balanced centered point ordered dithering. Uses $6 \times 6$ threshold matrix BALANCED_CENTERED_POINT_MAT.

DitherPunk.Bayer — Method

Bayer(level)

Ordered dithering using the Bayer matrix as a threshold matrix. The Bayer matrix is of dimension $2^{n+1} \times 2^{n+1}$, where $n$ is the level, which defaults to 1.

DitherPunk.CentralWhitePoint — Method

CentralWhitePoint()

Central white point ordered dithering. Uses $6 \times 6$ threshold matrix CENTRAL_WHITE_POINT_MAT.

DitherPunk.ClusteredDots — Method

ClusteredDots()

Clustered dots ordered dithering. Uses $6 \times 6$ threshold matrix CLUSTERED_DOTS_MAT.

DitherPunk.Rhombus — Method

Rhombus()

Diagonal ordered matrix with balanced centered points. Uses $8 \times 8$ threshold matrix RHOMBUS_MAT.

DitherPunk.bayer_matrix — Method

bayer_matrix(n::Int)::AbstractMatrix{Int}

Contruct (un-normalized) Bayer matrices of level n through recursive definition.

DitherPunk.closest_color — Method

Return color in ColorScheme cs that is closest to coloraccording to colordiff

DitherPunk.dither — Function

binarize([T::Type,] img, alg::AbstractDither, args...; kwargs...)

Dither img using algorithm alg.

Output

If no return type is specified, the default return type for binary dithering algorithms is Gray{Bool}. For color algorithms, the type of the colorscheme is used.

DitherPunk.dither! — Function

dither!([out,] img, alg::AbstractDither, args...; kwargs...)

Dither img using algorithm alg.

Output

If out is specified, it will be changed in place. Otherwise img will be changed in place.

DitherPunk.gradient_image — Method

Create linear grayscale gradient image of type Gray.

DitherPunk.linear2srgb — Method

Convert from linear to sRGB color space.

DitherPunk.srgb2linear — Method

Convert from sRGB to linear color space.

DitherPunk.test_on_gradient — Method

Test dithering algorithm alg on linear grayscale gradient from gradient_image and show stacked plot.

DitherPunk.upscale — Method

Upscale image by repeating individual pixels.