API Reference

DitherPunk.balanced_centered_point_dithering

DitherPunk.bayer_dithering

DitherPunk.bayer_matrix

DitherPunk.central_white_point_dithering

DitherPunk.clustered_dots_dithering

DitherPunk.gradient_image

DitherPunk.indices

DitherPunk.linear2srgb

DitherPunk.ordered_dithering

DitherPunk.print_braille

DitherPunk.rhombus_dithering

DitherPunk.show_dither

DitherPunk.srgb2linear

DitherPunk.test_on_gradient

DitherPunk.tile_matrix

DitherPunk.upscale

Functions

DitherPunk.balanced_centered_point_dithering — Method

balanced_centered_point_dithering(img; to_linear=false, invert_map=false)

Balanced centered point ordered dithering. Uses $6 \times 6$ threshold matrix BALANCED_CENTERED_POINT_MAT.

DitherPunk.bayer_dithering — Method

bayer_dithering(img::AbstractMatrix{<:Gray}; level=1, to_linear=false, invert_map=false)::BitMatrix

Ordered dithering using the Bayer matrix as a threshold matrix. The Bayer matrix is of dimension $2^{n+1} \times 2^{n+1}$, where $n$ is the level, which defaults to 1.

DitherPunk.bayer_matrix — Method

bayer_matrix(n::Int)::AbstractMatrix{Int}

Contruct (un-normalized) Bayer matrices of level n through recursive definition.

DitherPunk.central_white_point_dithering — Method

central_white_point_dithering(img; to_linear=false, invert_map=false)

Central white point ordered dithering. Uses $6 \times 6$ threshold matrix CENTRAL_WHITE_POINT_MAT.

DitherPunk.clustered_dots_dithering — Method

clustered_dots_dithering(img; to_linear=false, invert_map=false)

Clustered dots ordered dithering. Uses $6 \times 6$ threshold matrix CLUSTERED_DOTS_MAT.

DitherPunk.gradient_image — Method

Create linear grayscale gradient image of type Gray.

DitherPunk.indices — Method

Get indices from OffsetMatrix

DitherPunk.linear2srgb — Method

Convert from linear to sRGB color space.

DitherPunk.ordered_dithering — Method

ordered_dithering(img::AbstractMatrix{<:Gray}, mat::AbstractMatrix; invert_map=false)::BitMatrix

Ordered dithering using threshold maps. Takes a grayscale image img and a normalized threshold matrix mat. The threshold matrix is repeatedly tiled to match the size of img and is then applied as a per-pixel threshold map. Optionally, this final threshold map can be inverted by selecting invert_map=true.

DitherPunk.print_braille — Method

(Ab)use spy to interpret image as sparse array and print "sparsity pattern" in braille through UnicodePlots.

DitherPunk.rhombus_dithering — Method

rhombus_dithering(img; to_linear=false, invert_map=false)

Diagonal ordered matrix with balanced centered points, ordered dithering. Uses $8 \times 8$ threshold matrix.

DitherPunk.show_dither — Method

Apply Gray to indicate that image should be interpreted as a grayscale image. Optionally upscale image by integer value.

DitherPunk.srgb2linear — Method

Convert from sRGB to linear color space.

DitherPunk.test_on_gradient — Method

Test dithering algorithm alg on linear grayscale gradient from gradient_image and show stacked plot.

DitherPunk.tile_matrix — Method

tile_matrix(h, w, mat)

Repeatedly tile a smaller matrix mat to fill out an image of height h and width w.

DitherPunk.upscale — Method

Upscale image by repeating individual pixels.