API Reference

DitherPunk.balanced_centered_point_dithering

DitherPunk.bayer_dithering

DitherPunk.bayer_matrix

DitherPunk.central_white_point_dithering

DitherPunk.clustered_dots_dithering

DitherPunk.gradient_image

DitherPunk.linear2srgb

DitherPunk.ordered_dithering

DitherPunk.print_braille

DitherPunk.rhombus_dithering

DitherPunk.show_dither

DitherPunk.srgb2linear

DitherPunk.upscale

Functions

DitherPunk.balanced_centered_point_dithering — Method

Balanced centered point ordered dithering.

DitherPunk.bayer_dithering — Method

Ordered dithering using the Bayer matrix as a threshold map.

DitherPunk.bayer_matrix — Method

Contruct (un-normalized) Bayer matrices through recursive definition.

DitherPunk.central_white_point_dithering — Method

Central white point ordered dithering.

DitherPunk.clustered_dots_dithering — Method

Clustered dots ordered dithering.

DitherPunk.gradient_image — Method

Create grayscale gradient image of specified dimensions.

DitherPunk.linear2srgb — Method

Convert from linear to sRGB color space.

DitherPunk.ordered_dithering — Method

Ordered dithering using threshold maps. https://en.wikipedia.org/wiki/Ordereddithering#Pre-calculatedthreshold_maps

DitherPunk.print_braille — Method

(Ab)use spy to interpret image as sparse array and print "sparsity pattern" in braille through UnicodePlots.

DitherPunk.rhombus_dithering — Method

Diagonal ordered matrix with balanced centered points, ordered dithering.

DitherPunk.show_dither — Method

Apply Gray to indicate that image should be interpreted as a grayscale image. Optionally upscale image by integer value.

DitherPunk.srgb2linear — Method

Convert from sRGB to linear color space.

DitherPunk.upscale — Method

Upscale image by repeating individual pixels.